+++ examples/s4u/block-verification/monitor.cpp
@@ -8,12 +8,12 @@ Monitor::Monitor(std::vector<std::string> args) {
 
 void Monitor::operator()() {
   XBT_INFO("\nTime\t\tTotal network bytes produced\n");
-  double last_bytes_monitored = Node::network_bytes_produced;
+  long last_bytes_monitored = Node::network_bytes_produced;
   while (Node::active_nodes > 0) {
     if (last_bytes_monitored != Node::network_bytes_produced) {
         last_bytes_monitored = Node::network_bytes_produced;
         XBT_INFO(
-          "%f\t%d\n",
+          "%f\t%ld\n",
           simgrid::s4u::Engine::getClock(),
           Node::network_bytes_produced
         );
